make file operations (exists/delete/set_content/get_content/size) participate in transaction

+++ server/devpi_server/keyfs.py
@@ -104,7 +104,6 @@ class Filesystem:
 class FSWriter:
     def __init__(self, fs):
         self.fs = fs
-        self.pending_removes = []
         self.pending_renames = []
         self.changes = {}
 
@@ -126,6 +125,10 @@ class FSWriter:
         self.changes[relpath] = (typedkey.name, back_serial, value)
         self.pending_renames.append((tmp_path, target_path))
 
+    def record_rename_file(self, source, dest):
+        assert dest
+        self.pending_renames.append((source, dest))
+
     def __enter__(self):
         self.log = thread_push_log("fswriter%s:" % self.fs.next_serial)
         return self
@@ -133,16 +136,18 @@ class FSWriter:
     def __exit__(self, cls, val, tb):
         thread_pop_log("fswriter%s:" % self.fs.next_serial)
         if cls is None:
-            assert self.changes, "commit cannot be empty"
+            assert self.pending_renames, "commit cannot be empty"
             self.commit_to_filesystem()
             commit_serial = self.fs.next_serial - 1
-            if self.changes:
-                self.log.info("committed: %s", ",".join(self.changes))
-                self.fs._notify_on_commit(commit_serial)
+            size_basedir = len(str(self.fs.basedir))
+            changed = [x[1][size_basedir:] for x in self.pending_renames]
+            self.log.info("committed: %s", ",".join(changed))
+            self.fs._notify_on_commit(commit_serial)
         else:
             while self.pending_renames:
                 source, dest = self.pending_renames.pop()
-                os.remove(source)
+                if source is not None:
+                    os.remove(source)
             self.log.info("roll back at %s" %(self.fs.next_serial))
 
     def commit_to_filesystem(self):
@@ -153,14 +158,15 @@ class FSWriter:
                     self.fs.next_serial)
         self._direct_write(p, self.changes)
 
-        # do all renames and then removes
+        # do all renames and removes
         for source, dest in self.pending_renames:
-            rename(source, dest)
-        for dest in self.pending_removes:
-            try:
-                os.remove(dest)
-            except py.error.ENOENT:
-                pass
+            if source is None:
+                try:
+                    os.remove(dest)
+                except OSError:
+                    pass
+            else:
+                rename(source, dest)
 
         # finally increment the serial and write it out
         self.fs.next_serial += 1
@@ -290,7 +296,7 @@ class KeyFS(object):
                     fswriter.record_set(typedkey, val)
                     meth = self._import_subscriber.get(typedkey.name)
                     if meth is not None:
-                        meth(typedkey, val, back_serial)
+                        meth(fswriter, typedkey, val, back_serial)
 
     def subscribe_on_import(self, key, subscriber):
         assert key.name not in self._import_subscriber
@@ -490,6 +496,40 @@ class Transaction(object):
         self.at_serial = at_serial
         self.cache = {}
         self.dirty = set()
+        self.dirty_files = {}
+
+    def io_file_exists(self, path):
+        try:
+            return self.dirty_files[path] is not None
+        except KeyError:
+            return os.path.exists(path)
+
+    def io_file_set(self, path, content):
+        self.dirty_files[path] = content
+
+    def io_file_get(self, path):
+        try:
+            content = self.dirty_files[path]
+        except KeyError:
+            with open(path, "rb") as f:
+                return f.read()
+        if content is None:
+            raise IOError()
+        return content
+
+    def io_file_size(self, path):
+        try:
+            content = self.dirty_files[path]
+        except KeyError:
+            try:
+                return os.path.getsize(path)
+            except OSError:
+                return None
+        if content is not None:
+            return len(content)
+
+    def io_file_delete(self, path):
+        self.dirty_files[path] = None
 
     def exists_typed_state(self, typedkey):
         try:
@@ -537,7 +577,7 @@ class Transaction(object):
     def commit(self):
         if not self.write:
             return self._close()
-        if not self.dirty:
+        if not self.dirty and not self.dirty_files:
             threadlog.debug("nothing to commit, just closing tx")
             return self._close()
         try:
@@ -545,6 +585,14 @@ class Transaction(object):
                 for typedkey in self.dirty:
                     val = self.cache.get(typedkey)
                     fswriter.record_set(typedkey, val)
+                for path, content in self.dirty_files.items():
+                    if content is None:
+                        fswriter.record_rename_file(None, path)
+                    else:
+                        tmppath = path + "-tmp"
+                        with get_write_file_ensure_dir(tmppath) as f:
+                            f.write(content)
+                        fswriter.record_rename_file(tmppath, path)
                 commit_serial = fswriter.fs.next_serial
         finally:
             self._close()
@@ -575,3 +623,4 @@ def copy_if_mutable(val):
     elif isinstance(val, list):
         return list(val)
     return val
+
